+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog
 
+## v0.6.2 on t2s2-dev
+- fix more issues with open_end slots
+
+
 ## v0.6.1 on t2s2-dev
 - allow empty string for enddate if it's a slot with open end